requirements.txt
tensorflow
numpy
matplotlib
pandas
scikit-learn
pillow
